+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 **master**
+* Add `EveOnline::ESI::Models::MarketHistory`
+* Add `EveOnline::ESI::MarketHistory`
 
 **v0.13.0**